[feature][dingo-executor] Index reorg resolve confict
