+++ dingo-exec/src/main/java/io/dingodb/exec/operator/TxnScanWithCacheOpOperator.java
@@ -16,6 +16,7 @@
 
 package io.dingodb.exec.operator;
 
+import io.dingodb.common.log.LogUtils;
 import io.dingodb.exec.dag.Vertex;
 import io.dingodb.exec.operator.data.Context;
 import io.dingodb.exec.operator.params.ScanWithRelOpParam;
@@ -33,6 +34,10 @@ public final class TxnScanWithCacheOpOperator extends TxnScanWithRelOpOperatorBa
     @Override
     protected @NonNull Scanner getScanner(@NonNull Context context, @NonNull Vertex vertex) {
         ScanWithRelOpParam param = vertex.getParam();
+        if (context.getDistribution().getId().seq != 80012) {
+            int i = 0;
+            LogUtils.debug(log, "for debug");
+        }
         if (param.getCoprocessor(context.getDistribution().getId()) != null) {
             return RelOpUtils::doScan;
         }
